agm: fix incorrect pid of codec2 client
As codec2, directly doesn't interact with AgmIPC, still it was treated
as unique client. This is caused by using ABinder_getCallingPid which
provide the pid of binder transaction initiator.

As HAL and C2 don't use AGM IPC, and so for those callers always use
current pid (which is supposed to be HAL pid).

For any other clients use ABinder_getCallingPid.

Signed-off-by: Chiluka Rohith <rchiluka@qti.qualcomm.com>